refactor(sim): complete thermal_model_physics modular split (Issue #902)
Split the 2956-line monolithic src/sim/thermal_model_physics.rs into
focused submodules under a new src/sim/thermal_model_physics/
directory. The split unblocks the parent Issue #898 modular-split work
by removing the file/directory naming conflict that previously blocked
it.

Layout (all files in src/sim/thermal_model_physics/):
- mod.rs                  - submodule declarations + module docs
- hvac.rs                 - compute_zone_hvac_load (h_loss version, PR #929)
- batched_solver.rs       - solve_timesteps_batched
- solver_core.rs          - solve_timesteps, solve_timesteps_with_dt,
                            calculate_timestep_seconds,
                            estimate_time_constant_hours,
                            get_temperatures, get_hourly_temperatures,
                            calculate_analytical_loads, set_loads,
                            set_weather
- step_dispatcher.rs      - step_physics
- physics_impl.rs         - step_physics_5r1c, step_physics_6r2c,
                            step_physics_8r3c, step_physics_9r4c

Refactor reapplied on top of PR #928 (Vec-alloc reductions in
step_physics_5r1c) and PR #929 (h_loss-based compute_zone_hvac_load)
to avoid the rebase conflict that aborted attempt #0.

Implementation notes:
- All methods continue to be defined on the unified ThermalModel<T>
  type because Rust merges impl blocks from sibling modules within
  the same crate. The trait bound
    T: ContinuousTensor<f64> + From<VectorField> + AsRef<[f64]> + AsMut<[f64]>
  is repeated in every submodule so the methods all attach to the
  same ThermalModel<T> type.
- Cross-module callers use pub(crate) visibility
  (compute_zone_hvac_load, step_physics_5r1c/6r2c/8r3c/9r4c).
- #![allow(clippy::too_many_arguments)] on solve_timesteps_with_dt is
  preserved.
- compute_zone_hvac_load (h_loss version from PR #929) lives in
  hvac.rs as required by the issue.

Public API: unchanged. No external code imports from
crate::sim::thermal_model_physics::; all callers use method
invocations on ThermalModel<T>, which continue to resolve.

Verification:
- cargo check --lib: 0 errors
- cargo check --lib --all-features: 0 errors
- cargo test --release --lib -- --skip slow: 2463 passed, 2 ignored,
  1 filtered out
- cargo clippy --lib -- -D warnings: No issues found
- cargo fmt --check: clean
